// src/tools/src/util/toml/toml-enum.ts
/*
 * Copyright (c) Huawei Technologies Co., Ltd. 2026. All rights reserved.
 * This source file is part of the Cangjie project, licensed under Apache-2.0
 * with Runtime Library Exception.
 *
 * See https://cangjie-lang.cn/pages/LICENSE for license information.
 */

export const enum ObjType {
  DOT = 0,
  EXPLICIT = 1,
  ARRAY = 2,
  ARRAY_DOT = 3,
}

export enum LocationEnum {
  BEFORE = 0,
  AFTER = 1,
}